netstacklat: convert bpf_printk to optional debug feature
For production we need a way to disable any use of bpf_printk.

To track errors in production introduce a map for counting
these errors, as that will be exposed as a Prometheus counter
naming it netstacklat_errors_total.

The new "dbg" macro handled/hides if bpf_printk or counters
are enabled.

Signed-off-by: Jesper Dangaard Brouer <hawk@kernel.org>